fix(local-storage): make site usable without local storage
